backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-ratelimit==4.1.0
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
celery==5.3.4
django-extensions==3.2.3
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
dj-database-url